refactor(web): move utility helpers into services/_utils/ + tidy FRAGILE notes
K. Ten helper modules that aren't domain services were crowding
   apps/web/src/server/services/. Moved them under services/_utils/
   to keep the service tier index focused on the real domain services:
   fuzzy-score, mime-types, port-utils, upload-utils, github-graphql,
   sse-writer, jsonl-message-to-events, queued-message-store,
   dockview-layout-manager, terminal-layout-manager.

   Every importer (~16 files across server + tests + start-server)
   updated in lock-step. Biome reorganised imports on the touched files.

J. The two FRAGILE: ESM cycle leg comments in workspace-service.ts and
   task-service.ts referenced the deleted lib/workspace.ts shim. Rewrote
   them to describe the cycle's current shape (direct workspace-service
   ↔ task-service edge, with cronjob-service piggy-backing through
   task-service).

Refs: https://github.com/band-app/band/issues/535
