+++ apps/web/src/server/api/chats/router.ts
@@ -102,7 +102,10 @@ export const chatsRouter = t.router({
     // so the client can render a meaningful tab title without waiting for
     // a separate sessions.list. Subsequent reads are pure SQLite.
     if (workspace && (!chat.activeSessionId || chat.activeSessionSummary === undefined)) {
-      const resolved = await chatService.ensureActiveSessionSummary(input.chatId, workspace.worktree.path);
+      const resolved = await chatService.ensureActiveSessionSummary(
+        input.chatId,
+        workspace.worktree.path,
+      );
       if (resolved) {
         return { chat: resolved };
       }
