+++ apps/web/src/server/services/_utils/terminal-layout-manager.ts
